MaskImage
I coded this on my trip this morning but just got to where I had
internet.  This only includes the edges, I was going to start masking
the interior of our images as well, but I figure since we can’t keep
using photoshop, this should give us a good example for now.